==> Discogs/2.ipynb <==
import requests
import pandas as pd
import json
import time
import numpy as np

artistas = pd.read_csv('artistas.csv')

artistas

# REMPLAZAR FORMATOS y GENEROS

### NO FUNCIONA



formatos = {'album_format': ["'CD", "Album'", "EP'", "Comp'", "Single'", "'CD", "MiniAlbum", "Mixed'", "Comp", "Album", "Ltd'"]}
reemplazar_formatos = ["'CD", "Album'", "EP'", "Comp'", "Single'", "'CD", "MiniAlbum", "Mixed'", "Comp", "Album", "Ltd'"]
artistas['album_format'] = artistas['album_format'].replace(reemplazar_formatos, 'CD')

artistas['album_format']

## Vuelvo a intentarlo con regex

artistas.album_format = artistas.album_format.replace(".*CD.*", "CD", regex = True)

artistas.sample(10)

artistas.album_format.value_counts()

format_mapping = {
    '10"': 'Vinyl',
    '12"': 'Vinyl',
    '12", EP': 'Vinyl',
    '12", EP, Ltd': 'Vinyl',
    '10x12", Ltd, Red': 'Vinyl',
    '12", S/Edition': 'Vinyl',
    '12", Ltd, Promo, Smplr, W/Lbl': 'Vinyl',
    '12", W/Lbl': 'Vinyl',
    '12", S/Sided, Promo, TP': 'Vinyl',
    '12", Promo': 'Vinyl',
    '12", S/Sided, Promo': 'Vinyl',
    '2x12", Smplr': 'Vinyl',
    '12", S/Sided, Etch, Ltd': 'Vinyl',
    '2x12", W/Lbl': 'Vinyl',
    '12", Ltd, P/Mixed': 'Vinyl',
    '12", S/Sided': 'Vinyl',
    '12", Red': 'Vinyl',
    '12", Blu': 'Vinyl',
    '12", MiniAlbum': 'Vinyl',
    '12", Maxi': 'Vinyl',
    '20x12" + Box, Ltd, Num': 'Vinyl',
    '7"': 'Vinyl',
    '7", S/Sided, Single, Dlx, Num, Pic, RM, S/Edition': 'Vinyl',
    '2x12"': 'Vinyl',
    '12", MP, W/Lbl': 'Vinyl',
    '12", Ltd': 'Vinyl',
    'CD': 'CD',
    'Album': 'CD',
    'EP': 'CD',
    'Comp': 'CD',
    'Single': 'CD',
    'MiniAlbum': 'CD',
    'Mixed': 'CD',
    'Cass, Mixed': 'Cassette',
    'Cass, Comp': 'Cassette',
    'Cass, Promo, Smplr': 'Cassette',
    'File, MP3': 'Digital',
    'File, WAV': 'Digital',
    'DVD, NTSC, cod': 'DVD_VHS',
    'VHS, Advance, NTSC': 'DVD_VHS',
    'Acetate, 10"': 'Acetato',
    'Acetate, 12"': 'Acetato',
    'CDr, Mixed': 'CD',
    'CDr, Comp, Promo': 'CD',
    'CDr, Comp, Ltd': 'CD',
    '20x12" + Box, Ltd, Num': 'Vinyl',
    '0':'undefined'
}

# Ahora usamos el diccionario directamente en .map()
artistas['album_format_2'] = artistas['album_format'].map(format_mapping)
artistas['album_format_2'].unique()

artistas.album_format_2.value_counts()

len(artistas)

## Investigamos por qué hay tantos nulos

artistas.descripcion

artistas['descripcion'].value_counts()

## Crear lista de palabras clave.

keywords = [
    'pop', 'synthpop', 'electropop', 'dream pop', 'electronic', 'bubblegum pop', 'teen pop', 'indie pop',
    'rock', 'rock clásico', 'rock alternativo', 'hard rock', 'punk rock', 'garage rock', 'grunge',
    'glam rock', 'psychedelic rock', 'folk rock', 'prog rock', 'heavy metal', 'thrash metal', 'death metal',
    'black metal', 'power metal', 'doom metal', 'symphonic metal', 'nu metal', 'metalcore', 'hip-hop',
    'rap', 'old school hip-hop', 'trap', 'drill', 'lo-fi hip-hop', 'east coast hip-hop', 'west coast hip-hop',
    'gangsta rap', 'conscious rap', 'boom bap', 'r&b', 'soul', 'soul clásico', 'neo-soul', 'motown',
    'funk', 'contemporary r&b', 'quiet storm', 'electrónica', 'house', 'techno', 'trance', 'dubstep',
    'drum and bass', 'edm', 'synthwave', 'ambient', 'edm', 'vaporwave', 'reggae', 'reggae roots', 'dub',
    'ska', 'rocksteady', 'dancehall', 'ragga', 'latina', 'caribeña', 'salsa', 'cumbia', 'merengue',
    'bachata', 'reguetón', 'bolero', 'latin pop', 'ranchera', 'banda', 'música norteña', 'tango', 'country',
    'country clásico', 'country pop', 'bluegrass', 'americana', 'honky tonk', 'alt-country', 'jazz',
    'bebop', 'swing', 'cool jazz', 'jazz fusión', 'smooth jazz', 'free jazz', 'dixieland', 'hard bop',
    'blues', 'delta blues', 'chicago blues', 'blues rock', 'texas blues', 'country blues', 'electric blues',
    'folk', 'folk rock', 'folk pop', 'indie folk', 'folk tradicional', 'americana', 'nueva canción',
    'gospel', 'gospel tradicional', 'gospel contemporáneo', 'southern gospel', 'urban gospel', 'música clásica',
    'barroco', 'clásico', 'romanticismo', 'renacimiento', 'modernismo', 'neoclasicismo', 'minimalismo', 'ópera',
    'ópera barroca', 'ópera clásica', 'ópera romántica', 'ópera moderna', 'ópera buffa', 'ópera seria',
    'música contemporánea', 'serialismo', 'música electrónica académica', 'música aleatoria', 'música concreta',
    'música experimental', 'noise', 'avant-garde', 'drone', 'música ambiental experimental', 'música étnica',
    'música folclórica', 'música andina', 'música árabe', 'música africana', 'música india', 'música china',
    'flamenco', 'fado', 'balcan folk', 'música céltica', 'irish folk', 'scottish folk', 'folk bretaña',
    'música tradicional asiática', 'k-pop', 'j-pop', 'música tradicional japonesa', 'música tradicional china',
    'música tradicional tailandesa', 'world music', 'ethno jazz', 'worldbeat', 'afrobeat', 'bossa nova',
    'latin jazz', 'new age', 'ambient', 'meditación', 'música espiritual', 'healing music', 'música experimental',
    'noise', 'industrial', 'indie', 'alternativo', 'indie rock', 'indie pop', 'indie folk', 'shoegaze',
    'lo-fi', 'dream pop', 'chillout', 'lounge', 'trip-hop', 'future bass', 'jazz-dance']


## Crear una expresión regular a partir de las palabras clave

pattern = r'\b(?:' + '|'.join(keywords) + r')\b'

pattern

artistas['descripcion'] = artistas['descripcion'].str.lower()

## Filtrar y conservar solo las palabras clave


artistas['gender'] = artistas['descripcion'].str.findall(pattern)

artistas.sample(100)


## Queremos quitar listas, y creamos sets para que me desaparezcan las repetidas.

artistas['gender'] = artistas['gender'].apply(lambda x: set(x))

artistas.iloc[950].descripcion

artistas.info()

## Cambio la columna de genero a string para que me deje eliminar las llaves

artistas['gender']= artistas['gender'].astype(str)


artistas

artistas['gender'] = artistas['gender'].replace("set()", "").str.replace("{","").str.replace("}","").str.replace("'","")

artistas

artistas.to_csv('Remplazar_gender.csv', header=True, index= False)
