==> Discogs/3.ipynb <==
import requests
import pandas as pd
import json
import time
import numpy as np

formatos = pd.read_csv('Remplazar_gender.csv')

## Rellenamos no descripcion en las filas de genero que no contenga info


formatos['gender'] = formatos['gender'].fillna("no description")

formatos['gender']

## Quiero saber en la columna de album_format cuantos contienen 'album'

filtro_album = formatos[formatos['album_format'].str.contains('album', case=False, na=False)]

filtro_album

## Remplazo los albums los 0s por varios, ya que si no tenia info la columna era porque eran mas de un tipo de formato. No se podía hacer webscrapping

formatos['album_format'] = formatos['album_format'].replace("0", "varios")

formatos.sample(10)

formatos.head()

## Quiero que remplace los album por vinyl

formatos['album_format'] = formatos['album_format'].replace("album", "vinyl")

formatos

round(formatos['album_year'].mean())

## Queremos quitar en la columna de año de album los que pongan 0.

formatos.album_year.value_counts()

## Lo rellenamos por el año anterior

formatos['album_year'] = formatos['album_year'].replace(0, method='ffill')

formatos.album_year.value_counts()

formatos.to_csv('formatos.csv', header=True, index= False)
