==> Discogs/4.ipynb <==
import requests
import pandas as pd
import json
import time
import numpy as np

artistas_final = pd.read_csv('formatos.csv')

artistas_final

artistas_final['gender'].value_counts()

## Queremos borrar las dos columnas que no necesitamos

artistas_final = artistas_final.drop(columns=['uri', 'albums',])

artistas_final = artistas_final.drop(columns=['descripcion'])

artistas_final.head()

nuevo_orden = ['artist_id', 'name', 'type', 'album_name', 'gender','album_year','album_id', 'album_format', 'album_format_2']
artistas_final = artistas_final[nuevo_orden]

artistas_final.head()

artistas_final.to_csv('discogs.csv', header=True, index= False)

artistas_final_pivot = artistas_final.pivot_table(index=['name', 'album_name'], columns='gender', values='artist_id').reset_index()

artistas_final_pivot.sample()

artistas_final_index = artistas_final.set_index(['name', 'album_name'])

artistas_final_index
